requirements.txt
fastapi==0.75.2
sqlmodel==0.0.6
SQLAlchemy==1.4.35
uvicorn==0.17.6
gunicorn==20.1.0
pymysql==1.0.2
simplejson==3.17.6
python-decouple==3.6
loguru==0.6.0
unipath==1.1
PyJwt==2.3.0
passlib==1.7.4
pydantic==1.9.0
python-multipart==0.0.5
bcrypt==3.2.0
pandas==1.4.2
Crypto==1.4.1
pycryptodome==3.14.1
dataset==1.5.2
slugify==0.0.1
hypercorn[trio]==0.13.2
cryptography==36.0.2
python-arango==7.3.3
requests-toolbelt==0.9.1